+++ sensors_func.h
@@ -1,21 +1,18 @@
-//sensors.h
-//Autor: Maria Jarek
-//        Magdalena Zych
-//Data: 06.05.2020
+//sensors_func.h
+//Autorzy: Magdalena Zych
+//	   Maria Jarek
+//Data: 22.04.2020
 
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
 #include <ctype.h>
 #include <pthread.h>
-#include "sensor_function.h"
 #include <stdbool.h>
+#include "sensor_function.h"
 #include "parameters.h"
 
-bool check_arguments(char *arguments[]);
-void* diag_server_func(void* arg);
-bool check_param_communication(char str[]);
-void set_new_sleep_time(char action[]);
+struct sensor_parameters **parameters;
 
 struct sensor_threads_info
 {
@@ -28,3 +25,8 @@ union intInBuffer  //unia potrzebna do zapisania liczby typu int do bufora
   int intValue;
   char buffer[5];
 };
+
+bool check_sensors_arguments(char *arguments[]);
+void* diag_server_func(void* arg);
+bool check_param_communication(char str[]);
+void set_new_sleep_time(char action[]);
