+++ testowanie.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+#include "tests.h"
+
+
+int main()
+{
+	testuj();
+	return 0;
+}
