+++ tests.h
@@ -1,0 +1,11 @@
+//tests.h
+//Autorzy: Magdalena Zych
+//	   Krzysztof Gorlach
+//Data: 24.05.2020
+
+#include "control_func.h"
+#include "sensors_func.h"
+#include <CUnit/Basic.h>
+#include <CUnit/CUnit.h>
+
+void testuj();
